requirements.txt
google-cloud-bigquery==3.13.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.4
numpy==1.24.3
tqdm==4.66.1
python-dotenv==1.0.0
